requirements.txt
absl-py==2.1.0
altair==4.0.0
cffi==1.17.1
contourpy==1.3.0
cycler==0.12.1
flatbuffers==25.1.24
fonttools==4.55.7
jax==0.4.30
jaxlib==0.4.30
kiwisolver==1.4.7
matplotlib==3.9.4
mediapipe==0.10.20
ml_dtypes==0.5.1
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
pandas==1.5.3
pillow==11.1.0
protobuf==4.25.6
pycparser==2.22
pyparsing==3.2.1
scipy==1.13.1
sentencepiece==0.2.0
sounddevice==0.5.1
